shnurok.lab2/src/main/java/ru/ssau/tk/shnurok/lab2/operations/implementations/RightSteppingDifferentialOperator.java
package ru.ssau.tk.shnurok.lab2.operations.implementations;

import ru.ssau.tk.shnurok.lab2.functions.coredefenitions.MathFunction;
import ru.ssau.tk.shnurok.lab2.operations.abstractclasses.SteppingDifferentialOperator;

public class RightSteppingDifferentialOperator extends SteppingDifferentialOperator {

    public RightSteppingDifferentialOperator(double step) {
        super(step);
    }

    @Override
    public MathFunction derive(MathFunction function){
        return new MathFunction() {
            @Override
            public double apply(double x) {
                return (function.apply(x+step)-function.apply(x))/step;
            }
        };
    }
}
